+++ plugins/catalog-graphql/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @backstage/plugin-catalog-graphql
 
+## 0.3.16-next.0
+
+### Patch Changes
+
+- 3280711113: Updated dependency `msw` to `^0.49.0`.
+- Updated dependencies
+  - @backstage/types@1.0.2-next.0
+  - @backstage/catalog-model@1.1.4-next.0
+  - @backstage/config@1.0.5-next.0
+
 ## 0.3.15
 
 ### Patch Changes