+++ plugins/cost-insights/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @backstage/plugin-cost-insights
 
+## 0.12.1-next.0
+
+### Patch Changes
+
+- f9bbb3be37: Provide the ability to change the base currency from USD to any other currency in cost insights plugin
+- 3280711113: Updated dependency `msw` to `^0.49.0`.
+- Updated dependencies
+  - @backstage/core-components@0.12.1-next.0
+  - @backstage/core-plugin-api@1.1.1-next.0
+  - @backstage/plugin-catalog-react@1.2.2-next.0
+  - @backstage/catalog-model@1.1.4-next.0
+  - @backstage/config@1.0.5-next.0
+  - @backstage/theme@0.2.16
+  - @backstage/plugin-cost-insights-common@0.1.1
+
 ## 0.12.0
 
 ### Minor Changes