+++ plugins/example-todo-list-backend/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @internal/plugin-todo-list-backend
 
+## 1.0.8-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.16.1-next.0
+  - @backstage/plugin-auth-node@0.2.8-next.0
+  - @backstage/config@1.0.5-next.0
+  - @backstage/errors@1.1.4-next.0
+
 ## 1.0.7
 
 ### Patch Changes